backend/requirements.txt
pandas
mlflow
notebook
scikit-learn
numpy
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
fastapi
uvicorn
python-multipart
jinja2
keras
tensorflow
pytest
pytest-cov